+++ src/styles/colors.js
@@ -1,6 +1,6 @@
 export default {
     primary: '#00C2D4',
-    grey: '#EFF3F2',
+    grey: '#EEEEEE',
     white: '#FFFFFF',
     black: '#000000',
     transparentBlack: 'rgba(0, 0, 0, 0.5)',
